[Maps] rename file_upload to maps_file_upload (#89225)
* migrate file_upload plugin to maps_file_upload

* update plugins list

* results of running node scripts/build_plugin_list_docs

* fix build problems

* remove fileUpload from limits.yml
