+++ packages/kbn-optimizer/limits.yml
@@ -25,7 +25,6 @@ pageLoadAssetSize:
   esUiShared: 326798
   expressions: 224120
   features: 31211
-  fileUpload: 24717
   globalSearch: 43548
   globalSearchBar: 62888
   globalSearchProviders: 25554
@@ -106,3 +105,4 @@ pageLoadAssetSize:
   stackAlerts: 29684
   presentationUtil: 28545
   spacesOss: 18817
+  mapsFileUpload: 23775
