+++ x-pack/.i18nrc.json
@@ -20,7 +20,7 @@
     "xpack.endpoint": "plugins/endpoint",
     "xpack.enterpriseSearch": "plugins/enterprise_search",
     "xpack.features": "plugins/features",
-    "xpack.fileUpload": "plugins/file_upload",
+    "xpack.fileUpload": "plugins/maps_file_upload",
     "xpack.globalSearch": ["plugins/global_search"],
     "xpack.globalSearchBar": ["plugins/global_search_bar"],
     "xpack.graph": ["plugins/graph"],
