--- x-pack/plugins/file_upload/README.md
@@ -1,3 +1,0 @@
-# File upload
-
-Backend and core front-end react-components for GeoJson file upload. Only supports the Maps plugin.
