+++ x-pack/plugins/maps/kibana.json
@@ -8,7 +8,7 @@
     "features",
     "inspector",
     "data",
-    "fileUpload",
+    "mapsFileUpload",
     "uiActions",
     "navigation",
     "visualizations",
