+++ x-pack/plugins/maps_file_upload/README.md
@@ -1,0 +1,3 @@
+# Maps File upload
+
+Deprecated - plugin targeted for removal and will get merged into file_upload plugin
