+++ x-pack/plugins/maps_file_upload/common/constants/file_import.ts
