+++ x-pack/plugins/maps_file_upload/kibana.json
@@ -1,5 +1,5 @@
 {
-  "id": "fileUpload",
+  "id": "mapsFileUpload",
   "version": "8.0.0",
   "kibanaVersion": "kibana",
   "server": true,
