+++ x-pack/plugins/maps_file_upload/mappings.ts
