+++ x-pack/plugins/maps_file_upload/public/components/index_settings.js
