+++ x-pack/plugins/maps_file_upload/public/components/json_import_progress.js
