+++ x-pack/plugins/maps_file_upload/public/components/json_index_file_picker.js
