+++ x-pack/plugins/maps_file_upload/public/components/json_upload_and_parse.js
