+++ x-pack/plugins/maps_file_upload/public/get_file_upload_component.ts
