+++ x-pack/plugins/maps_file_upload/public/kibana_services.js
