+++ x-pack/plugins/maps_file_upload/public/plugin.ts
