+++ x-pack/plugins/maps_file_upload/public/util/file_parser.test.js
