+++ x-pack/plugins/maps_file_upload/public/util/geo_json_clean_and_validate.test.js
