+++ x-pack/plugins/maps_file_upload/public/util/http_service.js
