+++ x-pack/plugins/maps_file_upload/public/util/indexing_service.js
