+++ x-pack/plugins/maps_file_upload/public/util/size_limited_chunking.js
