+++ x-pack/plugins/maps_file_upload/server/kibana_server_services.js
