+++ x-pack/plugins/maps_file_upload/server/plugin.js
