+++ x-pack/plugins/maps_file_upload/server/routes/file_upload.js
