+++ x-pack/plugins/maps_file_upload/server/telemetry/file_upload_usage_collector.ts
