+++ x-pack/plugins/maps_file_upload/server/telemetry/mappings.ts
